Make it print data in float

+++ secllm_cpp/tensor.h
@@ -129,10 +129,10 @@ class Tensor {
         for (int b = 0; b < B; ++b) {
           for (int m = 0; m < M; ++m) {
             std::cout << "[" << b << ", " << m
-                      << ", 0, 0] = " << data_[b * M * K * N + m * K * N]
+                      << ", 0, 0] = " <<(float) data_[b * M * K * N + m * K * N]
                       << ", [" << b << ", " << m << ", " << K - 1 << ", "
                       << N - 1
-                      << "] = " << data_[b * M * K * N + m * K * N + K * N - 1]
+                      << "] = " <<(float) data_[b * M * K * N + m * K * N + K * N - 1]
                       << std::endl;
           }
         }
@@ -142,18 +142,18 @@ class Tensor {
         int N = shape_[2];
 
         for (int b = 0; b < B; ++b) {
-          std::cout << "[" << b << ", 0, 0] = " << data_[b * M * N] << ", ["
+          std::cout << "[" << b << ", 0, 0] = " <<(float) data_[b * M * N] << ", ["
                     << b << ", " << M - 1 << ", " << N - 1
-                    << "] = " << data_[b * M * N + M * N - 1] << std::endl;
+                    << "] = " <<(float) data_[b * M * N + M * N - 1] << std::endl;
         }
       } else if (dim == 2) {
         int M = shape_[0];
         int N = shape_[1];
 
-        std::cout << "[0, 0] = " << data_[0] << ", [" << M - 1 << ", " << N - 1
+        std::cout << "[0, 0] = " <<(float) data_[0] << ", [" << M - 1 << ", " << N - 1
                   << "] = " << data_[M * N - 1] << std::endl;
       } else if (dim == 1) {
-        std::cout << "[0] = " << data_[0] << ", [" << shape_[0] - 1
+        std::cout << "[0] = " <<(float) data_[0] << ", [" << shape_[0] - 1
                   << "] = " << data_[shape_[0] - 1] << std::endl;
       }
     }
